Error metadata specification change + languages
This commit changes the specification of error messages, so the contract
base class holds the errors in a list of dicts. This is beneficial so
the contract metadata can be read also without the error_collection
module.

The error_collection module still retains some of the underlying data
structure from the previous specification. This could perhaps be
changed, however the code is maybe 'good enough'.

NotImplementedWarnings:

 - There is still only support for 1 language!
 - Adding a second language may result in runtime errors.

The verify function checks that languages use a correct code and there
is also a test ready to check if languages implement a similar set of
implemented error results.